+++ Release-notes.md
@@ -29,8 +29,8 @@ be rather easily re-implemented (can you sense the irony?).
 that was nothing more than a wrapper class didn't feel relevant to
 me anymore. I hope you agree.
 
-- I have removed the RequiredMode class, which have caused some of
-the validation attributes to be modified.
+- I renamed the RequiredMode enum to UseSeparator then moved it to
+the Validation namespace, since it is only used there.
 
 
 NExtra.Demo